add Composite Primary Keys to TaskReleaseEntity
