+++ src/main/java/com/developing/simbir_product/entity/ReleaseEntity.java
@@ -1,6 +1,5 @@
 package com.developing.simbir_product.entity;
 
-
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
