+++ src/main/java/com/developing/simbir_product/entity/TaskReleaseEntity.java
@@ -1,62 +1,34 @@
 package com.developing.simbir_product.entity;
 
 import javax.persistence.Column;
+import javax.persistence.EmbeddedId;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 import javax.persistence.Table;
-import java.util.UUID;
 
 @Entity
 @Table(name = "TASK")
 public class TaskReleaseEntity {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private UUID id;
+    @EmbeddedId
+    private TaskReleaseId taskReleaseId;
 
-    @Column(name = "task_id")
-    private UUID taskId;
-
-    @Column( name = "release_id")
-    private UUID releaseId;
-
-    @Column(name = "isComplete")
+    @Column(name = "is_complete")
     private int isComplete;
 
     public TaskReleaseEntity() {
     }
 
-    public TaskReleaseEntity(UUID id, UUID taskId, UUID releaseId, int isComplete) {
-        this.id = id;
-        this.taskId = taskId;
-        this.releaseId = releaseId;
+    public TaskReleaseEntity(TaskReleaseId taskReleaseId, int isComplete) {
+        this.taskReleaseId = taskReleaseId;
         this.isComplete = isComplete;
     }
 
-    public UUID getId() {
-        return id;
+    public TaskReleaseId getTaskReleaseId() {
+        return taskReleaseId;
     }
 
-    public void setId(UUID id) {
-        this.id = id;
-    }
-
-    public UUID getTaskId() {
-        return taskId;
-    }
-
-    public void setTaskId(UUID taskId) {
-        this.taskId = taskId;
-    }
-
-    public UUID getReleaseId() {
-        return releaseId;
-    }
-
-    public void setReleaseId(UUID releaseId) {
-        this.releaseId = releaseId;
+    public void setTaskReleaseId(TaskReleaseId taskReleaseId) {
+        this.taskReleaseId = taskReleaseId;
     }
 
     public int getIsComplete() {
